refactor: don't splice array
